# mark: compile -k/-m expression once, reuse for all matches