@RonnyPfannschmidt 2020-05-11
With the revival of Mark expressions and keyword expressions as custom syntax the legacy status is gone, but we do pay a price for keeping mark and keyword expressions